Decouple pose prior from gps (#3065)
