+++ src/colmap/geometry/gps_test.cc
@@ -454,30 +454,5 @@ TEST(GPS, UTMToEllGRS80) {
   }
 }
 
-TEST(PosePrior, Equals) {
-  PosePrior prior;
-  prior.position = Eigen::Vector3d::Zero();
-  prior.position_covariance = Eigen::Matrix3d::Identity();
-  prior.coordinate_system = PosePrior::CoordinateSystem::CARTESIAN;
-  PosePrior other = prior;
-  EXPECT_EQ(prior, other);
-  prior.position.x() = 1;
-  EXPECT_NE(prior, other);
-  other.position.x() = 1;
-  EXPECT_EQ(prior, other);
-}
-
-TEST(PosePrior, Print) {
-  PosePrior prior;
-  prior.position = Eigen::Vector3d::Zero();
-  prior.position_covariance = Eigen::Matrix3d::Identity();
-  prior.coordinate_system = PosePrior::CoordinateSystem::CARTESIAN;
-  std::ostringstream stream;
-  stream << prior;
-  EXPECT_EQ(stream.str(),
-            "PosePrior(position=[0, 0, 0], position_covariance=[1, 0, 0, 0, 1, "
-            "0, 0, 0, 1], coordinate_system=CARTESIAN)");
-}
-
 }  // namespace
 }  // namespace colmap
