+++ src/colmap/scene/database.h
@@ -30,7 +30,7 @@
 #pragma once
 
 #include "colmap/feature/types.h"
-#include "colmap/geometry/gps.h"
+#include "colmap/geometry/pose_prior.h"
 #include "colmap/scene/camera.h"
 #include "colmap/scene/image.h"
 #include "colmap/scene/two_view_geometry.h"
